+++ Innowise.Music/Configuration/GoogleAuthenticationSettings.cs
@@ -1,0 +1,13 @@
+namespace Innowise.Music.Configuration;
+
+public class GoogleAuthenticationSettings
+{
+    public const string SectionName = "GoogleAuthentication";
+
+    public GoogleSettings? Google { get; set; }
+}
+
+public class GoogleSettings
+{
+    public string? ClientId { get; set; }
+}