+++ Innowise.MusicIdentityServer/Models/User/GoogleTokenDto.cs
@@ -1,0 +1,6 @@
+namespace Innowise.MusicIdentityServer.Models.User;
+
+public class GoogleTokenDto
+{
+    public string Token { get; set; }
+}